New config param: Fs123SignalFile
If non-empty, then when we catch a signal, we append a line to the file
named <Fs123SignalFile>.<getpid()>.  If Fs123SignalFile is relative, the
file will be opened in the cwd of the running process (see Fs123RunDir).

Default value:  "fs123signal"

Don't pretend that we have async-signal safe code in complaints.hpp
or log_channel.hpp.
Don't use complaints, log_channel or syslog in signal handlers.

call backtrace_symbol_fd in sighandler