Remove special handling of estalecookie=0
In the past, when the server set the estalecookie header to zero (or
didn't set it at all), the client interpreted that as a "don't care"
value.  Even if the e-s-c changed from zero to a non-zero value we
didn't return ESTALE.

But that special case broke when we added 'enhanced consistency', which
hashed the e-s-c into the ino - even if it was zero.  But we left
the special case in cookie_mismatch.

This patch removes the special case for estalecookie=0 in
cookie_mismatch.  Now, a zero value of estalecookie is no different from
any other value.

BUT - the estalecookie is only expected in /a, /d, and /f replies.  The
server typically omits the fs123-estalecookie header from other replies
(/x, /s, /l, /n).  Don't call cookie_mismatch for those replies.

Change-Id: I19bd5613ebb6c6303f6888db8025f1db4a45b05a