Maintain statistics about fifh->contents.size
Add a comment about possibly reducing the memory associated with large
open directories.

Add an (atomic) update_hwm method to stats.hpp.

Add an opendir_bytes_hwm statistic so we can decide if it's ever worth
the trouble to implement a more memory-frugal fhstate.

Move desres/docs/Notes.readdir to the public docs/ dir

Also (try to) clarify the comment in app_mount.cpp that
refers to it.

Change-Id: I2ec6f11aca342dfd66dcd36a2cd24566f9ec95a3