max-age==0 implies direct_io
Open makes an /a/ request to establish the initial value of
the 'validator' for new files.  (This is not new)

If the reply has max-age=0 then open sets the
fi->direct_io bit in the reply_open(..., fi) and it does *not*
add the file handle to the openfilemap.

Our expectation is that the kernel will forward all requests
directly to fuse without consulting its buffer cache, so there
is no need to regularly revalidate the file with the openfilemap
machinery.

Change-Id: I08ec276be70165486ba439ee0c356c773c36e3cf