Work around double-close bug in fuse_unmount
Change the order in fuse_teardown to:

    fs123_destroy (i.e., our ll_ops callback)
    fuse_unmount
    fuse_session_destroy (but don't call fs123_destroy)

By shutting down all of "our" stuff before calling fuse_unmount, we
are not confused by fuse_unmount's buggy double-close.

Refactor fuseful.cpp so all calls to fuse_unmount follow a consistent
pattern.

Keep some of the diagnostics that identified the double-close:

When an autoclose fails, include the numeric value of the file
descriptor (fileno(p) or dirfd(p)) in the what() message.

Clarify the requirements for a 'close_error_handler' in
autoclosers.hpp

Change-Id: I25b4d7a26a9c24efe527765892e9c31bc147f7f8