Add refcounted_scoped_timer, ut_timeit.cpp
refcounted_scoped_timer is in scoped_timer.hpp It's (roughly) the
std::chrono version of refcounted_scoped_nanotimer (which we've had
for a while.

Move the timing harnesses in ut_philox and ut_threefry to
a new ut_timeit.cpp.

Also add a new header: mt_timeit.hpp (which uses
refcounted_scoped_timer), but it's commented out because it produces
confusing results (in ut_timit.cpp).

Change-Id: I0c4b3448bbe225e102458323bdba5699c7d25b71