Remove --allow_unencrypted_replies options
Also remove the .NONE DO_NOT_ENCODE_SID logic.

It shouldn't be that easy to enable and then disable encryption.

New rule:
If you don't want encryption, don't configure a secret manager.

If the server is configured with a secret manager, then always
use it to encrypt replies.  Send a 406 to clients that don't present
the appropriate Accept-encoding.

If the client is configured with a secret manager, then reject any
replies that are not encrypted.

Also, fix --allow_unencrypted_requests.  It had been a no-op.
Now, if the server is configured with a secret manager and
if allow_unencrypted_requests is false (it defaults to true), the
server will reject non-/e-ncrpyted requests.

Fix the reply code-path in the server so that if there's a secret
manager, it encrypts even when errno is non-zero.  This is a bit
wasteful because such replies have an empty body, so there's nothing
to encrypt. But the code is cleaner on both ends.

Change-Id: I7a77702058233e4d2143e5e3ec80ff698e2b8e67