Add a new -oFs123Subprocess option
If the option is non-empty, the mount process starts an subprocess
as if by:
  system(Subprocess)

When system() returns, the mount is shut down and the fs123 process
exits.

If the mount is shut down by other means (umount, signal), then the
mount is removed from the system, but fs123p7 process itself does not
exit until the subprocess exits - thereby avoiding situations where
the subprocess and the parent process are "competing" for stdin.

Change-Id: Ibdbb3bc81beffa28752312d0893a613b1d1343c5