Make If-None-Match requests in the stale-while-reval window
We should see lots of If-None-Match requests and 304 replies when
objects in the diskcache fall in the the stale-while-revalidate
window.  We weren't because the diskcache code wasn't forwarding the
stale reply to the 'upstream' backend (the http backend).  Instead, it
was synthesizing an 'invalid' reply, which was perfecly safe, but
would never satisfy the http backend's requirements for generating an
If-None-Match header.

The fix is really only a couple of lines of plumbing, but it's in very
fragile code.

There's also a new unit-test to confirm that we're exercising the
INM/304 logic.

A few cosmetic changes slipped in as well.  The largest of which
is to eliminate include/diskcache_statistic_names in favor of
a macro in diskcache.hpp.

Change-Id: I2476f3994831864d6a17930c5678b5cb8935e9a2