Fix a serious bug when refreshing a block that fails invalidator-check
The 'retry' simply code read the wrong block because of a bad
cut-and-paste.  This bug caused a spate of segfaults after touching
actively used binary on the server.

The fix also adds some threeroe sums to a couple of DIAGs.

A regression test will be added in a subsequent patch.

Change-Id: Id21b4ba776720702868781049147d61252375b3d