Limit the amount of data written when updating a diskcache file
It's critical to not overwrite the threeroe_content, but while
were at it, we might as well write as little as possible.

This should prevent a fairly rare race where one thread does a full
replacement of a file in diskcache and another concurrently tries to
update just the cache-control metadata.

Change-Id: I14e848be7e3831084926766360a836a110c72eb3