Default to encrypting requests if sharedkeydir is non-empty
On the client, if -oFs123Sharedkeydir is non-empty, then
default to encrypting requests with /e.

In the server library, if the handler specifies a non-empty secret_manager,
then default to --allow_unencrypted_requests=false.

These defaults should be slightly less surprising when secretbox
is enabled.  They can easily be overridden.

Change-Id: Ia570436031b9b0042de70d5b632fdad997d96c95