Don't try to update the diskcache in-place
Experiments with multiple threads doing pwrite convinced us that Linux
doesn't guarantee the all-or-none atomicity of one thread pwrite-ing
into an open file while another thread reads from it.

So, don't do that.  Instead, when we have just a few bytes to change
in the "metadata" of an on-disk file, we still serialize the whole
thing with the usual open(tmpfile)/write/close/rename idiom.  A few
more bytes get sent to the kernel, but we avoid the torn writes.

Add a counter to record how many bytes are written to update 304s

Change-Id: I2d2047dd24d43bf156a56866c9c5d4e7e379d2ed