Conditionally use CURLOPT_CONNECT_TO or CURLOPT_RESOLVE
If LIBCURL_VERSION_NUM leads us to believe they'll "work".

Don't use the namecache with https if libcurl is so old that
we can't use CURLOPT_CONNECT_TO or CURLOPT_RESOLVE.

Move some TL;DR comments around.

And add a FIXME about the lifetime of the curl_slist
argument to CURLOPT_HTTPHEADER.

Change-Id: I5df2b63d1ca5360b778e1e7485d14a502b2d9a0f