Deep-six CacheFancySharing
The existing code was based on a misunderstanding about the atomicity
of reads and writes.

"Fixing" it would require tricky new logic and configuration options.
So this patch removes the option entirely and leaves some comments
behind with our current understanding of tradeoffs.

(Also update ut_expiring to use ut.hpp because it threw an error
when jenkins ran it even though  we've changed nothing even remotely
near it).

Change-Id: I6dca078b3291534a7a458ee0ebe07cd5968c31c1